% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expfit.R
\name{expfit}
\alias{expfit}
\title{Estimate Exponential Growth rate from  Aoristic data}
\usage{
expfit(
  x,
  niter = 1e+05,
  nburnin = 50000,
  thin = 10,
  nchains = 4,
  rPrior = "dnorm(mean=0,sd=0.05)",
  rSampler = NULL,
  parallel = FALSE,
  seeds = 1:4
)
}
\arguments{
\item{x}{A ProbMat class object}

\item{niter}{Number of MCMC iterations. Default is 500,000.}

\item{nburnin}{Mumber of iterations discarded for burnin. Default is 250,000.}

\item{thin}{Thinning interval}

\item{nchains}{Number of MCMC chains}

\item{rPrior}{A string defining prior for the growth parameter r. Default is 'dnorm(mean=0,sd=0.05)'.}

\item{rSampler}{A list containing settings for the MCMC sampler. Default is null and employs nimble's Default sampler (RW sampler).}

\item{parallel}{Logical specifying whether the chains should be run in parallel or not.}

\item{seeds}{Random seed for each chain. Default is 1:4.}
}
\description{
Fits an exponential growth model to \code{ProbMat} calss objects.
}
\details{
(To be completed)
}
