% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icarfit.R
\name{icarfit}
\alias{icarfit}
\title{Fits a random walk ICAR model to Aoristic data}
\usage{
icarfit(
  x,
  niter = 1e+05,
  nburnin = 50000,
  thin = 10,
  nchains = 4,
  sigmaPrior = "dexp(1)",
  sigmaSampler = NULL,
  parallel = FALSE,
  seeds = 1:4
)
}
\arguments{
\item{x}{A ProbMat class object}

\item{niter}{Number of MCMC iterations. Default is 500,000.}

\item{nburnin}{Mumber of iterations discarded for burnin. Default is 250,000.}

\item{thin}{Thinning interval}

\item{nchains}{Number of MCMC chains}

\item{sigmaPrior}{A string defining prior for the sigma parameter. Default is 'dexp(1)'.}

\item{sigmaSampler}{A list containing settings for the MCMC sampler. Default is null and employs nimble's Default sampler (RW sampler).}

\item{parallel}{Logical specifying whether the chains should be run in parallel or not.}

\item{seeds}{Random seed for each chain. Default is 1:4.}
}
\description{
Estimates parameters of a multinomial probability distribution that describes the shape of the of the time-frequency distribution of an observed sets of events with chronological uncertainty. The function is wrapper for fitting a 1D random walk ICAR model via nimble.
}
\details{
(To be completed)
}
