% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logisticfit.R
\name{logisticfit}
\alias{logisticfit}
\title{Fits a Logistic growth model on Aoristic data}
\usage{
logisticfit(
  x,
  niter = 1e+05,
  nburnin = 50000,
  thin = 10,
  nchains = 4,
  rPrior = "dexp(1/0.001)",
  mPrior = "dunif(1,z)",
  rSampler = NULL,
  mSampler = NULL,
  parallel = FALSE,
  seeds = 1:4
)
}
\arguments{
\item{x}{A ProbMat class object}

\item{niter}{Number of MCMC iterations. Default is 100,000.}

\item{nburnin}{Mumber of iterations discarded for burnin. Default is 50,000.}

\item{thin}{Thinning interval}

\item{nchains}{Number of MCMC chains}

\item{rPrior}{A string defining prior for the growth parameter r. Default is 'dexp(1/0.01)'.}

\item{mPrior}{A string defining prior for the point of maximum growth rate m. Default is 'dunif(1,z)', where 'z' is the number of time-blocks.}

\item{rSampler}{A list containing settings for the MCMC sampler for the parameter 'r'. Default is null and employs nimble's Default sampler (RW sampler).}

\item{mSampler}{A list containing settings for the MCMC sampler for the parameter 'm'. Default is null and employs nimble's Default sampler (RW sampler).}

\item{parallel}{Logical specifying whether the chains should be run in parallel or not.}

\item{seeds}{Random seed for each chain. Default is 1:4.}
}
\description{
Fits an exponential growth model to \code{ProbMat} calss objects.
}
\details{
(To be completed)
}
