% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotProbMat.R
\name{plot.probMat}
\alias{plot.probMat}
\title{Plot Aoristic Sums}
\usage{
\method{plot}{probMat}(
  x,
  xlab = NULL,
  ylab = NULL,
  type = "asum",
  calendar = "BP",
  lwd = 1,
  col = 1,
  minortick = NULL,
  add = FALSE,
  ...
)
}
\arguments{
\item{x}{\code{probMat} class object generated using the \code{generateProbMat()}.}

\item{xlab}{Label for the x-axis. Default based on \code{calendar}.}

\item{ylab}{Label for the y-axis. Default is 'Summed Probability' (if \code{type}='asum') or 'Probability Mass' (when \code{type}='dens').}

\item{type}{Either 'asum' for Aoristic Sum, 'dens' for probability mass. Default is 'asum'.}

\item{calendar}{Either \code{'BP'} or \code{'BCAD'}. Indicate whether the x-axis should be displayed in BP or BC/AD. Default is  \code{'BP'}.}

\item{lwd}{Line width. Default is 1.}

\item{col}{Line col. Default is 'black'}

\item{minortick}{Interval for minor ticks in the x-axis label. Default is estimated based on timescale}

\item{add}{if set to \code{TRUE} adds the line and point graph on existing plot.}

\item{...}{Additional arguments affecting the plot.}
}
\value{
No return value (plot function)
}
\description{
Plot summed probabilities of aoristic weights.
}
