% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sampledf}
\alias{sampledf}
\title{Sample aoristic data (data.frame)}
\format{
A data.frame class object with two columns defining the start and the end of each even (\code{sample.df})
}
\usage{
sampledf
}
\description{
Sample datasets to illustrate data formats required for \code{createProbMat()}.
}
\examples{
data(sampledf)
x  <- createProbMat(x=sampledf,timeRange = c(6500,4001),resolution= 100)
}
\keyword{datasets}
