% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{samplemat}
\alias{samplemat}
\title{Sample aoristic data (matrix)}
\format{
A matrix class object storing the probability of each event (row) in each time-block (column)
}
\usage{
samplemat
}
\description{
Sample datasets to illustrate data formats required for \code{createProbMat()}.
}
\examples{
data(samplemat)
x <- createProbMat(pmat=samplemat,timeRange = c(5000,3001),resolution=100)
plot(x)
}
\keyword{datasets}
